test(frontend): consolidate activity-timeline wire-boundary coverage (PR #41 b99cea79)
Reproduce PR #41's activity-timeline wire-boundary test on PR #42: drive raw
snake_case TF turn.summary frames through normalizeTFEvent -> dispatchSSEEvent
(the exact path useSSE uses), asserting a native activity_label yields a nonblank
group heading and a labelless step update never erases it.

Fold in the native-only, no-synthetic-label cases (a blank step activity_label is
also "no opinion"; a leading step stays blank instead of fabricating a label from
its title; a following activity update then opens a properly-labeled group) and
rename the earlier turn-summary.test.ts into this file — it drove the same wire
path, so this removes the duplicated coverage the reviewer flagged while retaining
every no-synthetic-label expectation.

Copilot-Session: 0a90e0a4-4933-4f98-8c37-e6f8846c902b
